refactor(hooks): split agent_briefing under the 300-line cap, honest source note (#400, #401)
Two follow-ups on the #400 fix, both raised by review.

1. The `# source:` citation on `_NON_SPECIALIST_META_AGENTS` named
   `~/.claude/agents/dispatch.md` "in this repo". That file ships in no
   repository — it is a user-scope Claude Code agent file on the reader's
   own machine, so the citation was unverifiable by any other contributor
   or by CI, which is precisely what §8 forbids. It is replaced by an
   honest statement of what the constant is: a convention of the consumer
   environment, whose *consequence* is verifiable and pinned by tests.
   A design decision is not dressed up as a sourced empirical fact.

2. Filtering a reserved meta-agent name was silent. Since the whole defect
   class here is "the briefing is silently inactive", the drop is now
   logged, so a future name collision is diagnosable instead of mysterious.

`agent_briefing.py` was over the local 300-line cap before this work
(452 lines) and the fix pushed it to 476. Keyword extraction, the PG
connect/query pair and the stderr logger move to sibling modules,
re-exported via `__all__`; the module is back to 297 lines and #401 is
closed by construction rather than deferred. Event gating and the
specialist-roster loader stay put — the loader reads the `CLAUDE_DIR`
attribute the test suite monkeypatches on this module.

Gates, all re-run by hand: the #400 regression test still fails on the
pre-fix source with the reported symptom and passes after; full suite
7243 passed / 142 skipped / 0 failed in deterministic order; ruff check
and ruff format --check green. One order-dependent failure surfaced under
pytest-randomly, outside this change's blast radius and filed with its
evidence as #402.
